# documentation/api/OPENAPI-3.0.yaml
openapi: 3.0.0
info:
  title: Sugar-Daddy Platform - Comprehensive API
  description: |
    完整的 Sugar-Daddy 平台 API 文檔
    包含所有 81 個端點的詳細說明
    
    ### 系統組成
    - 🔐 Auth Service (11 個端點)
    - 🎬 Content-Streaming Service (11 個端點)  
    - 💳 Payment Service (19 個端點)
    - 🎁 Subscription Service (10 個端點)
    - 🤖 Recommendation Service (13 個端點)
    - 🌉 API Gateway (1 個端點)
    
    **總計: 81 個 API 端點**
  version: 1.0.0
  contact:
    name: Sugar-Daddy Backend Team
    email: backend@sugardaddy.com
  license:
    name: Proprietary
    url: https://sugardaddy.com/license
servers:
  - url: http://localhost:3000
    description: API Gateway (本地開發)
    variables:
      port:
        default: '3000'
  - url: https://api.sugardaddy.com/v1
    description: 生產環境
  - url: https://staging-api.sugardaddy.com/v1
    description: 預發佈環境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token 認證
        登錄後獲得 token，在 Authorization header 中使用
        格式: Authorization: Bearer <token>
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Bad Request
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - statusCode
        - message

    # ==================== Auth Service Schemas ====================
...
    Recommendation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        score:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
          example: Based on your viewing history
...
  responses:
    UnauthorizedError:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
            errors:
              - field: token
                message: Invalid or expired token
    NotFoundError:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not Found
            errors:
              - message: Resource not found
    BadRequestError:
      description: 請求格式錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/recommendations:
    get:
      tags:
        - Recommendation
      summary: 獲取推薦列表
      description: 獲取當前用戶的個性化推薦內容
      operationId: getRecommendations
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 推薦內容列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Recommendation
      summary: 創建推薦配置
      description: 為用戶創建自定義推薦配置
      operationId: createRecommendationConfig
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferences:
                  type: array
                  items:
                    type: string
                excludeWatched:
                  type: boolean
                  default: true
      responses:
        '201':
          description: 推薦配置已創建
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/recommendations/{contentId}:
    get:
      tags:
        - Recommendation
      summary: 獲取相似內容推薦
      description: 獲取與指定內容相似的推薦
      operationId: getSimilarContent
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 相似內容推薦
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/recommendations/interactions:
    post:
      tags:
        - Recommendation
      summary: 記錄用戶交互
      description: 記錄用戶與內容的交互（評論、分享等）
      operationId: recordInteraction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contentId:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [view, like, share, comment]
                metadata:
                  type: object
              required:
                - contentId
                - type
      responses:
        '201':
          description: 交互已記錄
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/recommendations/refresh/{userId}:
    post:
      tags:
        - Recommendation
      summary: 刷新用戶推薦
      description: 強制刷新用戶的推薦算法
      operationId: refreshUserRecommendations
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 推薦已刷新
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/recommendations/update-scores:
    post:
      tags:
        - Recommendation
      summary: 更新推薦評分
      description: 後臺任務：更新所有用戶的推薦評分
      operationId: updateRecommendationScores
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 評分更新已啟動
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/recommendations/clear-cache:
    post:
      tags:
        - Recommendation
      summary: 清除推薦快取
      description: 清除推薦引擎的快取
      operationId: clearRecommendationCache
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 快取已清除
        '401':
          $ref: '#/components/responses/UnauthorizedError'
security:
  - BearerAuth: []